# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: واجهة دليل موردي المخابز والحلويات
servers:
  - url: /api
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      parameters:
        - name: categoryId
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
        - name: city
          in: query
          required: false
          schema: { type: string, maxLength: 80 }
        - name: q
          in: query
          required: false
          description: Search request title, details, category, business name, or city
          schema: { type: string, maxLength: 100 }
      responses:
        "200":
          description: Requests visible to the authenticated buyer or supplier
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Request" }
        "401": { description: Login required }
        "403": { description: Account is not allowed to view requests }
    post:
      operationId: createRequest
      tags: [requests]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RequestInput" }
      responses:
        "201":
          description: Request published
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Request" }
        "400": { description: Invalid request data or city is unavailable }
        "401": { description: Buyer login required }
        "403": { description: Only active business owners may publish requests }
        "404": { description: Category is unavailable }
...
  /buyer/register:
    post:
      operationId: registerBuyer
      tags: [buyer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyerRegisterInput"
      responses:
        "201":
          description: Buyer account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuyerAuthResponse"
        "400": { description: Invalid registration }
        "409": { description: Phone or email already used }
  /buyer/login:
    post:
      operationId: loginBuyer
      tags: [buyer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyerLoginInput"
      responses:
        "200":
          description: Buyer session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuyerAuthResponse"
        "401": { description: Invalid credentials }
...
  /buyer/me:
    get:
      operationId: getBuyerMe
      tags: [buyer]
      responses:
        "200":
          description: Current buyer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuyerMeResponse"
        "401": { description: Buyer authentication required }
  /buyer/contact:
    post:
      operationId: createBuyerContactLog
      tags: [buyer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyerContactInput"
      responses:
        "201":
          description: Contact log created and WhatsApp URL prepared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuyerContactResponse"
        "400": { description: Invalid contact input }
        "401": { description: Buyer authentication required }
        "404": { description: Supplier not found }
...
components:
  schemas:
...
    Request:
      type: object
      required: [id, categoryId, categoryName, title, description, quantity, unit, frequency, city, businessName, status, createdAt, expiresAt]
      properties:
        id: { type: integer }
        categoryId: { type: integer }
        categoryName: { type: string }
        title: { type: string }
        description: { type: string }
        quantity: { type: number }
        unit: { type: string, enum: ["كيلو", "كرتون", "كيس", "علبة"] }
        frequency: { type: string, enum: ["مرة واحدة", "أسبوعي", "شهري"] }
        city: { type: string }
        businessName: { type: ["string", "null"] }
        status: { type: string, enum: ["active", "closed", "expired"] }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
...
    RequestInput:
      type: object
      required: [categoryId, quantity, unit, frequency, city, description]
      properties:
        categoryId: { type: integer, minimum: 1 }
        quantity: { type: number, exclusiveMinimum: 0 }
        unit: { type: string, enum: ["كيلو", "كرتون", "كيس", "علبة"] }
        frequency: { type: string, enum: ["مرة واحدة", "أسبوعي", "شهري"] }
        city: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 1000 }
...
    BuyerUser:
      type: object
      required: [id, fullName, phone, email, city, businessType, businessName, otherBusinessType, isOwner, jobTitle, createdAt, lastLogin]
      properties:
        id: { type: integer }
        fullName: { type: string }
        phone: { type: string }
        email: { type: [string, "null"], format: email }
        city: { type: string }
        businessType: { type: string }
        businessName: { type: [string, "null"] }
        otherBusinessType: { type: [string, "null"] }
        isOwner: { type: boolean }
        jobTitle: { type: [string, "null"] }
        createdAt: { type: string }
        lastLogin: { type: [string, "null"] }
    BuyerRegisterInput:
      type: object
      required: [fullName, phone, city, businessType, isOwner, password]
      properties:
        fullName: { type: string, minLength: 2, maxLength: 80 }
        phone: { type: string }
        email: { type: string, format: email, maxLength: 160 }
        city: { type: string }
        businessType: { type: string, enum: ["مخبز", "محل حلويات", "مخبز وحلويات", "كافيه", "مطعم", "أسرة منتجة", "أسر منتجة", "فندق", "آخر"] }
        businessName: { type: string, maxLength: 120 }
        otherBusinessType: { type: string, maxLength: 80 }
        isOwner: { type: boolean }
        jobTitle: { type: string, maxLength: 80 }
        password: { type: string, minLength: 6, maxLength: 128, pattern: "^[A-Za-z0-9]+$" }
        newsletterWeekly: { type: boolean }
        buyersGroup: { type: boolean }
...
    BuyerLoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier: { type: string }
        password: { type: string, minLength: 1 }
        rememberMe: { type: boolean, default: false }
...
    BuyerAuthResponse:
      type: object
      required: [success, message, user]
      properties:
        success: { type: boolean }
        requestCode: { type: string }
        message: { type: string }
        user: { $ref: "#/components/schemas/BuyerUser" }
    BuyerMeResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/BuyerUser" }
    BuyerContactInput:
      type: object
      required: [supplierId]
      properties:
        supplierId: { type: integer }
        message: { type: string, maxLength: 1000 }
    BuyerContactResponse:
      type: object
      required: [success, messageId, message, whatsappUrl]
      properties:
        success: { type: boolean }
        messageId: { type: string }
        message: { type: string }
        whatsappUrl: { type: string, format: uri }
...